backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.34.0
torch==2.1.0
streamlit==1.26.0
requests==2.31.0
numpy==1.27.0
pydantic==2.8.0
